int -> short

+++ src/main/java/cn/nukkit/level/format/generic/BaseFullChunk.java
@@ -39,7 +39,7 @@ public abstract class BaseFullChunk implements FullChunk, ChunkManager {
      */
     protected byte[] biomes;
 
-    protected int[] blocks;
+    protected short[] blocks;
 
     protected byte[] data;
 
@@ -434,7 +434,7 @@ public abstract class BaseFullChunk implements FullChunk, ChunkManager {
     }
 
     @Override
-    public int[] getBlockIdArray() {
+    public short[] getBlockIdArray() {
         return this.blocks;
     }
